sublimeapp/requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.4.3
Django==5.0.3
django-filter==24.2
iniconfig==2.0.0
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
pytest==8.1.1
sqlparse==0.4.4
tzdata==2024.1
django-widget-tweaks==1.5.0
python-decouple==3.8
factory-boy==3.3.0
Faker==23.3.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
stripe==9.1.0
django-filter